# docs/swagger.yaml
basePath: /
definitions:
...
  handler.RefundRequest:
    properties:
      reason:
        example: Product defective
        type: string
    type: object
...
  model.Order:
    properties:
      address:
        type: string
      created_at:
        type: string
      customer_name:
        type: string
      customer_phone:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.OrderItem'
        type: array
      payment_method:
        type: string
      shipping_method:
        type: string
      status:
        type: string
      total_price:
        type: number
      user_id:
        type: integer
    type: object
  model.OrderItem:
    properties:
      price:
        type: number
      product_id:
        type: integer
      product_name:
        type: string
      quantity:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Clothing Store.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Clothing Store API
  version: "1.0"
paths:
...
  /my/orders:
    get:
      consumes:
      - application/json
      description: Get a paginated list of orders for the currently logged-in user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List my orders
      tags:
      - orders
  /orders:
    post:
      consumes:
      - application/json
      description: Create a new order with items
      parameters:
      - description: Order Data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/model.Order'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Order'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - orders
  /orders/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an order by ID (only pending or cancelled)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: Order deleted successfully
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete an order
      tags:
      - orders
  /orders/{id}/approve:
    put:
      consumes:
      - application/json
      description: Change order status from 'chờ xử lý' to 'đã xử lý' (Staff/Admin
        only)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: Order approved successfully
          schema:
            type: string
        "400":
          description: Invalid ID or status
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Approve an order
      tags:
      - orders
  /orders/{id}/cancel:
    put:
      consumes:
      - application/json
      description: Cancel an order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: Order cancelled successfully
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Cancel an order
      tags:
      - orders
  /orders/{id}/pay:
    put:
      consumes:
      - application/json
      description: Mark an order as paid
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: Order paid successfully
          schema:
            type: string
        "400":
          description: Invalid ID
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Pay an order
      tags:
      - orders
  /orders/{id}/refund:
    put:
      consumes:
      - application/json
      description: Refund an order with a reason
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      - description: Refund Reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RefundRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: Order refunded successfully
          schema:
            type: string
        "400":
          description: Invalid ID or request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Refund an order
      tags:
      - orders
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"